#23 - directory scan wip
