+++ src/server/maupod/cmd/audioscan/main.go
@@ -1,20 +1,16 @@
 package main
 
 import (
-	"context"
-	"database/sql"
+	"errors"
 	"log"
 	"os"
-	"time"
+	"os/signal"
 
-	_ "github.com/lib/pq"
 	"github.com/mauleyzaola/maupod/src/server/pkg/data"
-	"github.com/mauleyzaola/maupod/src/server/pkg/data/orm"
-	"github.com/mauleyzaola/maupod/src/server/pkg/filemgmt"
-	"github.com/mauleyzaola/maupod/src/server/pkg/filters"
-	"github.com/mauleyzaola/maupod/src/server/pkg/media"
-	"github.com/mauleyzaola/maupod/src/server/pkg/pb"
+	"github.com/mauleyzaola/maupod/src/server/pkg/helpers"
 	"github.com/mauleyzaola/maupod/src/server/pkg/rule"
+	"github.com/mauleyzaola/maupod/src/server/pkg/simplelog"
+	"github.com/mauleyzaola/maupod/src/server/pkg/types"
 	"github.com/spf13/viper"
 )
 
@@ -27,7 +23,6 @@ func main() {
 }
 
 func init() {
-	log.SetFlags(log.Lshortfile | log.Ltime | log.Ldate)
 	viper.AddConfigPath(".")
 	viper.SetConfigType("yaml")
 	viper.SetConfigName(".maupod")
@@ -46,164 +41,56 @@ func run() error {
 		return err
 	}
 
-	var roots []string
-	var fileSystemStores = rule.ConfigurationFileSystemStores(config)
-	for _, v := range fileSystemStores {
-		roots = append(roots, v.Location)
-	}
-	for _, root := range roots {
-		if _, err = os.Stat(root); err != nil {
+	// create directory if not exists
+	imageStore := rule.ConfigurationFirstImageStore(config)
+	if imageStore != nil {
+		if err = os.MkdirAll(imageStore.Location, os.ModePerm); err != nil {
 			return err
 		}
+	} else {
+		return errors.New("could not find any image store in configuration")
 	}
 
-	scanDate := time.Now()
-
-	var db *sql.DB
-	if db, err = data.DbBootstrap(config); err != nil {
-		return err
-	}
-	defer func() {
-		if err = db.Close(); err != nil {
-			log.Println(err)
-		}
-	}()
-
-	store := &data.MediaStore{}
-	ctx := context.Background()
-	conn := db
-
-	var filter = filters.MediaFilter{}
-	var allMedia data.Medias
-	if allMedia, err = store.List(ctx, conn, filter, nil); err != nil {
+	nc, err := helpers.ConnectNATS()
+	if err != nil {
 		return err
 	}
 
-	mediaLocationKeys := allMedia.ToMap()
-	var cols = orm.MediumColumns
-	var fields = []string{
-		cols.Sha,
-		cols.FileExtension,
-		cols.Format,
-		cols.FileSize,
-		cols.Duration,
-		cols.OverallBitRateMode,
-		cols.OverallBitRate,
-		cols.StreamSize,
-		cols.Album,
-		cols.Track,
-		cols.Title,
-		cols.TrackPosition,
-		cols.Performer,
-		cols.Genre,
-		cols.RecordedDate,
-		cols.Comment,
-		cols.Channels,
-		cols.ChannelPositions,
-		cols.ChannelLayout,
-		cols.SamplingRate,
-		cols.SamplingCount,
-		cols.BitDepth,
-		cols.CompressionMode,
-		cols.EncodedLibrary,
-		cols.EncodedLibraryName,
-		cols.EncodedLibraryVersion,
-		cols.BitRateMode,
-		cols.BitRate,
-		cols.LastScan,
-		cols.ModifiedDate,
-		cols.TrackNameTotal,
-		cols.AlbumPerformer,
-		cols.AudioCount,
-		cols.BitDepthString,
-		cols.CommercialName,
-		cols.CompleteName,
-		cols.CountOfAudioStreams,
-		cols.EncodedLibraryDate,
-		cols.FileName,
-		cols.FolderName,
-		cols.FormatInfo,
-		cols.FormatURL,
-		cols.InternetMediaType,
-		cols.KindOfStream,
-		cols.Part,
-		cols.PartTotal,
-		cols.StreamIdentifier,
-		cols.WritingLibrary,
-		cols.Composer,
+	var logger types.Logger
+	logger = &simplelog.Log{}
+	logger.Init()
+
+	db, err := data.DbBootstrap(config)
+	if err != nil {
+		return err
 	}
 
-	insertFn := func(ctx context.Context, filename string, info *media.MediaInfo) error {
-		fileInfo, err := os.Stat(filename)
-		if err != nil {
-			return err
-		}
-		m := rule.NewMediaFile(info, filename, scanDate, fileInfo)
-
-		// if the location is the same and we made it here, that means we need to update the row
-		if val, ok := mediaLocationKeys[filename]; ok {
-			m.Id = val.Id
-			return store.Update(ctx, conn, m, fields...)
-		} else {
-			return store.Insert(ctx, conn, m)
-		}
+	var hnd types.Broker
+	hnd = NewMsgHandler(config, db, logger, nc)
+	if err = hnd.Register(); err != nil {
+		return err
 	}
 
-	for _, root := range roots {
-		if err = ScanFiles(ctx, root, config, mediaLocationKeys, insertFn); err != nil {
-			return err
+	// handle interruptions and cleanup resources
+	signalChan := make(chan os.Signal, 1)
+	cleanupDone := make(chan bool)
+	signal.Notify(signalChan, os.Interrupt)
+	go func() {
+		cleanup := func() {
+			hnd.Close()
+			cleanupDone <- true
 		}
-	}
-
-	return nil
-
-}
-
-func ScanFiles(ctx context.Context, root string, config *pb.Configuration,
-	mediaLocationKeys map[string]*pb.Media,
-	insertFn func(ctx context.Context, filename string, info *media.MediaInfo) error) error {
-	var err error
-	var files []string
-	start := time.Now()
-
-	walker := func(filename string, isDir bool) bool {
-		if isDir {
-			return false
-		}
-		if !rule.FileIsValidExtension(config, filename) {
-			return false
-		}
-
-		// a bit of speed improvement, avoid a second time scanning the same file unless it has been changed in the file system
-		if val, ok := mediaLocationKeys[filename]; ok {
-			if !rule.NeedsUpdate(val) {
-				return false
+	loop:
+		for {
+			select {
+			case <-signalChan:
+				break loop
 			}
 		}
+		cleanup()
+	}()
 
-		files = append(files, filename)
-		return false
-	}
-
-	log.Println("[DEBUG] started scanning")
-	if err = filemgmt.WalkFiles(root, walker); err != nil {
-		log.Println(err)
-		return err
-	}
-	log.Printf("[DEBUG] finished scanning %d files\n", len(files))
-
-	for _, f := range files {
-		info, err := media.InfoFromFile(f)
-		if err != nil {
-			log.Printf("[ERROR] cannot get mediainfo from file: %s %s\n", f, err)
-			continue
-		}
-		if err = insertFn(ctx, f, info); err != nil {
-			return err
-		}
-	}
-
-	log.Printf("[INFO] files: %d  elapsed: %s\n", len(files), time.Since(start))
+	<-cleanupDone
 
 	return nil
 }
