+++ src/server/proto/messages.proto
@@ -5,11 +5,17 @@ import "google/protobuf/timestamp.proto";
 import "media.proto";
 
 enum Message {
-    MESSAGE_ARTWORK_EXTRACT = 0;
+    MESSAGE_ARTWORK_SCAN = 0;
+    MESSAGE_AUDIO_SCAN = 1;
 }
 
 
 message ArtworkExtractInput {
     Media media = 1;
+    google.protobuf.Timestamp scan_date = 2;
+}
+
+message ScanDirectoryAudioFilesInput {
+    string root = 1;
     google.protobuf.Timestamp scan_date = 2;
 }
